Move API Tracing to debug module
