+++ docs/developer-guide/configuring.md
@@ -6,8 +6,8 @@ luma.gl is currently published as a single package, however there are still some
 
 | Optional Import            | Description |
 | ---                        | --- |
+| `import '@luma.gl/debug';`  | Install optional WebGL debug support. Enables creation of debug contexts. Import before creating contexts. |
 | `import 'luma.gl/webgl1';` | Install optional support for running luma.gl on WebGL1-only browsers. Import before creating contexts. |
-| `import 'luma.gl/debug';`  | Install optional WebGL debug support. Enables creation of debug contexts. Import before creating contexts. |
 | `import GL from 'luma.gl/constants';` | Import static WebGL constant definitions |
