+++ modules/core/debug.js
@@ -1,2 +1,4 @@
+// Deprecated import file, just issues warning
 // eslint-disable-next-line
-module.exports = require('./dist/es5/debug');
+console.warn(`importing 'luma.gl/debug' no longer has any effect. \
+Please import the separate '@luma.gl/debug' module instead`);
