+++ modules/core/src/multipass/canvas.js
@@ -1,5 +1,3 @@
-import 'luma.gl/debug';
-
 import {
   createGLContext,
   Model,
