+++ modules/debug/src/glsl-to-js-compiler/compile-shader.js
