+++ modules/debug/src/glsl-to-js-compiler/draw-model.js
