+++ modules/debug/src/webgl-api-tracing/parameter-definitions.js
