+++ modules/debug/src/webgl-api-tracing/webgl-debugger.js
@@ -1,0 +1,5 @@
+// Expose Khronos Debug support module on global context
+/* global window, global */
+const WebGLDebug = require('webgl-debug');
+const global_ = typeof global !== 'undefined' ? global : window;
+global_.WebGLDebug = WebGLDebug;
